+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.1.1] - 2023-04-13
+
+### Changed
+
+* Updated the bucket policy on the DataplaneLogsBucket bucket(s) to grant access to the logging service principal (logging.s3.amazonaws.com) for access log delivery
+
 ## [5.1.0] - 2023-02-06
 
 ### Added